Add creating classes only for teacher login.

+++ server/src/services/schoolClassService.tsx
@@ -1,8 +1,11 @@
 import prisma from "../config/prismaClient.js";
-import type {ClassTypes} from "../controllers/schoolClassController.js";
+import type { ClassTypes } from "../controllers/schoolClassController.js";
 
-export const getClassesService = async () => {
+export const getClassesService = async (teacherId: number) => {
     return prisma.schoolClass.findMany({
+        where: {
+            teacherId,
+        },
         include: {
             _count: {
                 select: {
@@ -16,18 +19,18 @@ export const getClassesService = async () => {
     });
 };
 
-export const createClassService = async ({
-                                             name,
-                                             type,
-                                         }: ClassTypes) => {
+export const createClassService = async (
+    teacherId: number,
+    {
+        name,
+        type,
+    }: ClassTypes
+) => {
     return prisma.schoolClass.create({
         data: {
             name,
             type,
-
-            // TODO:
-            // replace this with the logged-in teacher id
-            teacherId: 1,
+            teacherId,
         },
         include: {
             _count: {
@@ -40,9 +43,24 @@ export const createClassService = async ({
 };
 
 export const updateClassService = async (
+    teacherId: number,
     classId: number,
-    { name, type }: ClassTypes
+    {
+        name,
+        type,
+    }: ClassTypes
 ) => {
+    const schoolClass = await prisma.schoolClass.findFirst({
+        where: {
+            id: classId,
+            teacherId,
+        },
+    });
+
+    if (!schoolClass) {
+        throw new Error("Class not found");
+    }
+
     return prisma.schoolClass.update({
         where: {
             id: classId,
@@ -61,7 +79,21 @@ export const updateClassService = async (
     });
 };
 
-export const deleteClassService = async (classId: number) => {
+export const deleteClassService = async (
+    teacherId: number,
+    classId: number
+) => {
+    const schoolClass = await prisma.schoolClass.findFirst({
+        where: {
+            id: classId,
+            teacherId,
+        },
+    });
+
+    if (!schoolClass) {
+        throw new Error("Class not found");
+    }
+
     return prisma.schoolClass.delete({
         where: {
             id: classId,
@@ -69,10 +101,14 @@ export const deleteClassService = async (classId: number) => {
     });
 };
 
-export const getClassService = async (classId: number) => {
-    return prisma.schoolClass.findUnique({
+export const getClassService = async (
+    teacherId: number,
+    classId: number
+) => {
+    return prisma.schoolClass.findFirst({
         where: {
             id: classId,
+            teacherId,
         },
         include: {
             sus: {
